refactor: move info tab methods
